+++ package.json
@@ -9,7 +9,7 @@
         "lint": "eslint -c .eslintrc.js --ext .ts .",
         "test": "mocha --parallel",
         "testNoE2E": "npm run test -- --exclude **/e2eTest.ts",
-        "testData": "npm run test -- --exclude !packages/azure-kusto-data/**",
+        "testData": "npm run test -- --exclude !packages/azure-kusto-ingest/test/e2eTests/**",
         "testIngest": "npm run test -- --exclude !packages/azure-kusto-ingest/**",
         "testIngestNoE2E": "npm run testNoE2E -- --exclude !packages/azure-kusto-ingest/**",
         "testPipeline": "nyc --reporter lcovonly mocha --reporter mocha-junit-reporter --reporter-options outputs=true",
@@ -32,7 +32,6 @@
         "nyc": "^15.1.0",
         "prettier": "^2.7.1",
         "sinon": "^14.0.0",
-        "ts-node": "^10.9.1",
         "typescript": "^4.7.4",
         "@types/node": "^14.0.0"
     },
