+++ packages/azure-kusto-data/.npmrc
@@ -1,0 +1,4 @@
+registry=https://msazure.pkgs.visualstudio.com/One/_packaging/Azure-Kusto-Service/npm/registry/ 
+                        
+always-auth=true
+engine-strict=true
