--- packages/azure-kusto-data/index.ts
@@ -1,9 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-import KustoClient from "./source/client";
-import ClientRequestProperties from "./source/clientRequestProperties";
-import KustoConnectionStringBuilder from "./source/connectionBuilder";
-import * as KustoDataErrors from "./source/errors";
-
-export { KustoClient as Client, ClientRequestProperties, KustoConnectionStringBuilder, KustoDataErrors };
