+++ packages/azure-kusto-data/package.json
@@ -2,8 +2,8 @@
     "name": "azure-kusto-data",
     "version": "3.4.2",
     "description": "Azure Data Explorer Query SDK",
-    "main": "index.js",
-    "types": "index.d.ts",
+    "main": "src/index.js",
+    "types": "src/index.d.ts",
     "scripts": {
         "build": "tsc"
     },
@@ -29,25 +29,47 @@
         "directory": "azure-kusto-data"
     },
     "files": [
-        "source/**/*.js",
-        "source/**/*.d.ts",
-        "index.js",
-        "index.d.ts",
+        "src/**/*.js",
+        "srcn/**/*.d.ts",
+        "src.js",
+        "src/src.js",
+        "src/index.d.ts",
         "tsconfig.tsbuildinfo",
         "example.js"
     ],
     "author": "",
     "license": "ISC",
     "homepage": "https://github.com/Azure/azure-kusto-node/blob/master/azure-kusto-data/README.md",
+    "optionalDependencies": {
+        "@azure/msal-browser": "^2.28.1",
+        "@azure/msal-node": "^1.12.1"
+    },
+    "browser": {
+        "fs": false,
+        "os": false,
+        "process": false,
+        "./dist-esm/src/tokenProvider.js": "./dist-esm/src/tokenProvider.browser.js"
+    },
+    "devDependencies": {
+        "@types/mocha": "^9.1.1",
+        "@types/webpack-dev-server": "^4.7.2",
+        "ts-loader": "^9.3.1",
+        "webpack-cli": "^4.10.0"
+    },
     "dependencies": {
         "@azure/identity": "^2.1.0",
-        "@azure/msal-node": "^1.12.1",
+        "@azure/msal-common": "^7.3.0",
         "@types/node": "^18.6.4",
         "@types/uuid": "^8.3.4",
+        "assert": "^2.0.0",
         "axios": "^0.27.2",
+        "browsify": "^0.0.4",
+        "crypto-browserify": "^3.12.0",
         "follow-redirects": "^1.15.1",
+        "https-browserify": "^1.0.0",
         "moment": "^2.29.4",
-        "ts-node": "^10.9.1",
+        "stream-browserify": "^3.0.0",
+        "stream-http": "^3.2.0",
         "uuid": "^8.3.2"
     },
     "gitHead": "f8a5dae26d6d2ca2ab8b95953bb9b88a02e8e35d"
