+++ packages/azure-kusto-data/src/client.ts
