+++ packages/azure-kusto-data/src/clientRequestProperties.ts
