+++ packages/azure-kusto-data/src/cloudSettings.ts
