+++ packages/azure-kusto-data/src/errors.ts
