+++ packages/azure-kusto-data/src/index.ts
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import KustoClient from "./client";
+import ClientRequestProperties from "./clientRequestProperties";
+import KustoConnectionStringBuilder from "./connectionBuilder";
+import * as KustoDataErrors from "./errors";
+
+const kcsb = KustoConnectionStringBuilder.withUserPrompt("https://devdevon.westeurope.dev.kusto.windows.net");
+const cli = new KustoClient(kcsb);
+cli.execute("fast","TestTable2").then(resp=>console.log(resp))
+
+export { KustoClient as Client, ClientRequestProperties, KustoConnectionStringBuilder, KustoDataErrors };
