+++ packages/azure-kusto-data/src/response.ts
