+++ packages/azure-kusto-data/src/typeUtilts.ts
