+++ packages/azure-kusto-data/test/clientTest.ts
@@ -6,17 +6,17 @@ import assert from "assert";
 import { v4 as uuidv4 } from "uuid";
 import moment from "moment";
 
-import { KustoClient } from "../source/client";
+import { KustoClient } from "../src/client";
 
-import { ClientRequestProperties } from "../source/clientRequestProperties";
-import { KustoResponseDataSetV1, KustoResponseDataSetV2 } from "../source/response";
+import { ClientRequestProperties } from "../src/clientRequestProperties";
+import { KustoResponseDataSetV1, KustoResponseDataSetV2 } from "../src/response";
 
 import v2Response from "./data/response/v2.json";
 import v2ResponseError from "./data/response/v2error.json";
 import v1Response from "./data/response/v1.json";
 import v1_2Response from "./data/response/v1_2.json";
 import { Readable } from "stream";
-import ConnectionBuilder from "../source/connectionBuilder";
+import ConnectionBuilder from "../src/connectionBuilder";
 
 enum ExecutionType {
     Mgmt = "mgmt",
