+++ packages/azure-kusto-data/test/responseTest.ts
@@ -4,7 +4,7 @@
 import assert from "assert";
 
 import v2Response from "./data/response/v2.json";
-import { KustoResponseDataSetV2, V2Frames } from "../source/response";
+import { KustoResponseDataSetV2, V2Frames } from "../src/response";
 
 describe("KustoResultDataSet", () => {
     describe("#constructor()", () => {
