+++ packages/azure-kusto-data/tsconfig.json
@@ -1,5 +1,5 @@
 {
     "extends": "../../tsconfig.json",
-    "include": ["./source", "./test", "index.ts"],
+    "include": ["src", "src/*.ts","./test", "src/index.ts","webpack.config.ts"],
     "exclude": ["node_modules", "*.d.ts", "**/*.d.ts"]
 }
