+++ packages/quick_start/source/Utils.ts
@@ -1,14 +1,14 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import KustoConnectionStringBuilder from "azure-kusto-data/source/connectionBuilder";
+import KustoConnectionStringBuilder from "azure-kusto-data/src/connectionBuilder";
 import fs from "fs";
-import ClientRequestProperties from "azure-kusto-data/source/clientRequestProperties";
+import ClientRequestProperties from "azure-kusto-data/src/clientRequestProperties";
 import { v4 as uuidv4 } from "uuid";
 import { DataFormat } from "azure-kusto-ingest";
 import IngestionProperties, { dataFormatMappingKind } from "azure-kusto-ingest/source/ingestionProperties";
 import Console from "console";
-import KustoClient from "azure-kusto-data/source/client";
+import KustoClient from "azure-kusto-data/src/client";
 import IngestClient from "azure-kusto-ingest/source/ingestClient";
 import { BlobDescriptor, FileDescriptor } from "azure-kusto-ingest/source/descriptors";
 import { AuthenticationModeOptions } from "./SampleApp";
